Split relay dispatcher layers
