+++ config/max-lines-baseline.txt
@@ -31,7 +31,6 @@ inline src/main/ssh/ssh-relay-deploy.ts
 inline src/main/ssh/ssh-relay-session.ts
 inline src/main/updater.ts
 inline src/preload/index.ts
-inline src/relay/dispatcher.ts
 inline src/relay/git-handler.ts
 inline src/relay/pty-handler.ts
 inline src/renderer/src/components/LinearItemDrawer.tsx
